table and sugg orginizing, styling
