+++ grantly/src/styles/suggestionStyles.js
@@ -42,72 +42,9 @@ export const suggestionStyles = makeStyles(theme => ({
     fontFamily: "Nunito Sans",
     fontWeight: "600"
   },
-  // suggestionWrapper: {
-  //   display: 'flex',
-  //   flexDirection: 'column',
-  //   margin: "1em",
-  // },
   userSuggestion: {
     fontFamily: "Nunito Sans",
     fontSize: '1.5em',
     width: '8.25em',
-  },
-  suggestionLi: {
-    display: 'flex',
-    marginTop: '.5em',
-    padding: '13px',
-    borderTop: '3px solid #77D4D0',
-  },
-  p_wrapper: {
-    width: '400px'
-  },
-  suggestionParagraph: {
-    width: '50%',
-    overflow: "auto",
-    borderBottom: '3px solid #77D4D0',
-    padding: '10px',
-    marginTop: '5px',
-    fontSize: '1.11rem'
-  },
-  suggestionHeader: {
-    width: '50%',
-    paddingLeft: '10px',
-    paddingBottom: '5px',
-    marginTop: '5px',
-    fontSize: '1.3rem'
-  },
-  suggestionNone: {
-    width: '20%',
-    borderBottom: '3px solid #77D4D0',
-    padding: '10px',
-    marginTop: '5px',
-    fontSize: '1.25rem'
-  },
-  delSuggestBtn: {
-    borderRadius: '50px',
-    alignSelf: 'center',
-    backgroundColor: '#EF7B5C',
-    boxShadow: 'none',
-    borderWidth: '0px',
-    margin: '2px 5px 2px 0',
-    padding: '6px',
-    height: '40px',
-    width: '40px',
-    "&:hover": {
-      backgroundColor: '#f0a692'
-    }
-  },
-  iconBtnWithSuggestions: {
-    backgroundColor: '#3DB8B3',
-    paddingRight: "5px",
-    border: "none"
-  },
-  iconBtnWithOutSuggestions: {
-    backgroundColor: 'none',
-    color: "000",
-    cursor: "unset",
-    "&:hover":{
-      backgroundColor: 'white'
-    }
   },
 }));
